Improved devFunctions and "Execute As"
