+++ package.json
@@ -19,8 +19,12 @@
     "discord.js": "^14.7.1",
     "dotenv": "^16.0.3",
     "image-converter": "^1.0.0",
+    "jimp": "^0.22.10",
     "mongoose": "^6.8.3",
-    "pureimage": "^0.3.15"
+    "pureimage": "^0.4.13"
   },
-  "description": ""
+  "description": "",
+  "devDependencies": {
+    "typescript": "^5.3.3"
+  }
 }
