+++ src/config.json
@@ -3,5 +3,10 @@
   "ROLL_COOLDOWN": 14400000,
   "TOKENS_PER_ROLL": 5,
   "TOKEN_PER_SACRIFICE": 1,
-  "LOCK_DURATION": 60000
+  "LOCK_DURATION": 60000,
+  "DISCORD_BOT_DEVS": [
+    "396233379890200579", //Le_Roux-nard
+    "253227788834832384",  //Eldras
+    "224130483888324608"  //Grizfreak
+  ]
 }
