+++ tsconfig.json
@@ -97,10 +97,5 @@
 		/* Completeness */
 		// "skipDefaultLibCheck": true,                      /* Skip type checking .d.ts files that are included with TypeScript. */
 		"skipLibCheck": true /* Skip type checking all .d.ts files. */,
-		"plugins": [
-			{
-				"transform": "ts-transformer-keys/transformer"
-			}
-		]
 	}
 }
